+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Analyzer: percent analyze
 - Analyzer: katakana and hiragana percent
+### Fixed
+- Analyzer: description characters count
 
 ## [0.5.9-19] - 2020-07-05
 ### Added