+++ CHANGELOG.md
@@ -5,9 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.5.9-20] - 2020-07-06
 ### Added
 - Analyzer: percent analyze
 - Analyzer: katakana and hiragana percent
+- Analyzer: frequency analyze
 ### Fixed
 - Analyzer: description characters count
 
@@ -581,7 +584,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Example story as usage.
 - Output story as markdown.
 
-[Unreleased]: https://github.com/nagisc007/storybuilder/compare/v0.5.9-19...HEAD
+[Unreleased]: https://github.com/nagisc007/storybuilder/compare/v0.5.9-20...HEAD
+[0.5.9-20]: https://github.com/nagisc007/storybuilder/releases/v0.5.9-20
 [0.5.9-19]: https://github.com/nagisc007/storybuilder/releases/v0.5.9-19
 [0.5.9-18]: https://github.com/nagisc007/storybuilder/releases/v0.5.9-18
 [0.5.9-17]: https://github.com/nagisc007/storybuilder/releases/v0.5.9-17